+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/userinfo/UserInfoManager.kt
@@ -14,6 +14,7 @@
 
 package de.netid.mobile.sdk.userinfo
 
+import android.net.Uri
 import de.netid.mobile.sdk.api.NetIdError
 import de.netid.mobile.sdk.model.UserInfo
 import de.netid.mobile.sdk.webservice.UserInfoCallback
@@ -21,10 +22,10 @@ import de.netid.mobile.sdk.webservice.WebserviceApi
 
 class UserInfoManager(private val listener: UserInfoManagerListener) {
 
-    fun fetchUserInfo(host: String, accessToken: String) {
+    fun fetchUserInfo(userinfoEndpoint: Uri, accessToken: String) {
         WebserviceApi.performUserInfoRequest(
             accessToken,
-            host,
+            userinfoEndpoint,
             object : UserInfoCallback {
                 override fun onUserInfoFetched(userInfo: UserInfo) {
                     listener.onUserInfoFetched(userInfo)