+++ CHANGELOG.md
@@ -36,6 +36,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed soft/hard to permission/login
 - Make subclaims optional, too
 - Broker/host is no longer configurable from the outside
+- Changes in calling verified app links
+- Use userinfo endpoint from discovery document
+- Get rid of copy of id token
+- Cleaned up unused string resources
+- Small refactoring about claims handling
 
 ### Fixed
 - Fixed UserInfo json parser error